make conformer and rotation dtype adjustable

+++ src/rfm_docking/dataset.py
@@ -55,7 +55,7 @@ def process_one_(args):
         U[:, -1] *= -1  # Correct for reflection if necessary
         R = U @ Vt
 
-    R = torch.tensor(R, dtype=torch.float32)
+    R = torch.tensor(R, dtype=torch.float64)
 
     inv_lattice = torch.inverse(
         torch.tensor(lattice_matrix_target, dtype=torch.float32)
@@ -63,11 +63,9 @@ def process_one_(args):
 
     smiles = row.ligand  # or .smiles
 
-    conformer_cart = smiles_to_pos(smiles, forcefield="mmff", device="cpu")
-
-    # conformer to fractional coordinates
-    # conformer_frac = conformer @ inv_lattice.T
-    # conformer_frac = conformer_frac.numpy()
+    conformer_cart = smiles_to_pos(
+        smiles, forcefield="mmff", device="cpu", dtype=torch.float64
+    )
 
     loading = int(row.loading)
     node_feats, edge_feats, edge_index = featurize_osda(smiles)
